Apply all remaining audit suggestions: validation, security, edge cases
WHBinding validation (all 7 fields):
- Cross-check 'who' matches actor, 'where' matches target
- Reject precedence escalation (actor claims higher than actual)
- 7 new tests for WHBinding cross-checks

PhiTraversalLog:
- Add is_all_passed() computed method (source of truth)
- finalize() synchronizes stored field with computed value

Governance:
- BFT FinalityMode::validate() enforces quorum_threshold >= 1
- EmergencyPolicy::validate() enforces activation > deactivation
- Deactivate variant stores normal_max_txs (no more hardcoded 1000)

Contract execution:
- Authorization check: actor governance_level must meet contract requirement
- Preconditions enforced: reject if any law not in tx.preconditions
- Postconditions enforced: reject if any law not in tx.postconditions
- Method name non-empty validation, args size limit (1MB)
- verify_contract_id() helper to check hash matches content
- 4 new tests: authorized, unauthorized, missing precondition, step limit

Security hardening:
- SymbolAddress MAX_SYMBOL_ADDRESS_LEN (4KB) constant
- Economics: reject negative tension_budget in effective_fee
- SCCE: cap activated symbols at 16, scan limit per symbol (1000)
- SCCE: select_relevant_subgraph returns u64 (not u32)
- Validator round-robin: sort by node_id for deterministic ordering
- total_txs accumulator widened from u32 to u64

93 tests passing. Clippy clean.

+++ crates/sccgub-node/src/main.rs
@@ -228,7 +228,7 @@ fn cmd_status(data_dir: &std::path::Path) {
     }
 
     let latest = blocks.last().unwrap();
-    let total_txs: u32 = blocks.iter().map(|b| b.body.transition_count).sum();
+    let total_txs: u64 = blocks.iter().map(|b| b.body.transition_count as u64).sum();
 
     println!("=== SCCGUB Chain Status ===");
     println!("  Data dir:       {:?}", data_dir);
